requirements.txt
aerich==0.8.0
aiofiles==23.2.1
aiogram==3.0.0b3
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.9.1
aiolimiter==1.2.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.3.0
APScheduler==3.11.0
asgiref==3.8.1
asyncclick==8.1.7.2
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dacite==1.8.1
dataclass-factory==2.16
defusedxml==0.7.1
dictdiffer==0.9.0
distro==1.9.0
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
fake-useragent==2.0.3
fastapi==0.115.6
fasteners==0.19
fonttools==4.55.3
frozenlist==1.5.0
get-video-properties==0.1.1
google==3.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
iso8601==1.1.0
jiter==0.8.2
kiwisolver==1.4.7
magic-filter==1.0.12
matplotlib==3.9.3
multidict==6.1.0
numpy==2.1.2
onlinesimru==2.0.4
openai==1.57.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pendulum==3.0.0
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.19
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pypika-tortoise==0.1.6
Pyrogram==2.0.106
pyshorteners==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
selenium==4.27.1
shazamio==0.7.0
shazamio_core==1.0.7
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.41.3
StrEnum==0.4.15
Telethon==1.38.1
time-machine==2.16.0
tomlkit==0.13.2
tortoise-orm==0.19.0
tqdm==4.67.1
transliterate==1.10.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
youthon==0.4.3
youtube-transcript-api==0.6.3
yt-dlp==2025.1.26
